Refactor proto paths and update build scripts
