+++ .gitignore
@@ -33,4 +33,5 @@ report.[0-9]_.[0-9]_.[0-9]_.[0-9]_.json
 # Finder (MacOS) folder config
 .DS_Store
 build
-proto/**
+/proto/**
+src/assets/proto/**
