+++ proto/node.ts
@@ -1,21 +1,21 @@
 import type * as grpc from '@grpc/grpc-js';
 import type { EnumTypeDefinition, MessageTypeDefinition } from '@grpc/proto-loader';
 
-import type { CexServiceClient as _cexBroker_CexServiceClient, CexServiceDefinition as _cexBroker_CexServiceDefinition } from './cexBroker/CexService';
+import type { cex_serviceClient as _cex_broker_cex_serviceClient, cex_serviceDefinition as _cex_broker_cex_serviceDefinition } from './cex_broker/cex_service';
 
 type SubtypeConstructor<Constructor extends new (...args: any) => any, Subtype> = {
   new(...args: ConstructorParameters<Constructor>): Subtype;
 };
 
 export interface ProtoGrpcType {
-  cexBroker: {
+  cex_broker: {
     Action: EnumTypeDefinition
     ActionRequest: MessageTypeDefinition
     ActionResponse: MessageTypeDefinition
-    CexService: SubtypeConstructor<typeof grpc.Client, _cexBroker_CexServiceClient> & { service: _cexBroker_CexServiceDefinition }
     SubscribeRequest: MessageTypeDefinition
     SubscribeResponse: MessageTypeDefinition
     SubscriptionType: EnumTypeDefinition
+    cex_service: SubtypeConstructor<typeof grpc.Client, _cex_broker_cex_serviceClient> & { service: _cex_broker_cex_serviceDefinition }
   }
 }
