+++ test/start-broker.test.ts
@@ -1,4 +1,4 @@
-import { describe, test, expect, beforeEach } from "bun:test";
+import { describe, test, expect } from "bun:test";
 import { startBrokerCommand } from "../src/commands/start-broker";
 
 describe("Start Broker Command", () => {
